upd: add cloudservice for DB, changed user-connection
