+++ Our-Project/package.json
@@ -22,6 +22,7 @@
 	"type": "module",
 	"dependencies": {
 		"axios": "^1.6.7",
+		"jwt-decode": "^4.0.0",
 		"valibot": "^0.30.0"
 	}
 }
