+++ Our-Project/src/lib/Axios/authAxios.ts
@@ -1,4 +1,6 @@
 import axios from "axios";
+import { jwtDecode } from "jwt-decode";
+
 import type { loginDTO, registerDTO } from "../DTO/auth/authDTO";
 import type { ResLoginDTO, ResRegisterDTO } from "../DTO/auth/authResponse";
 
@@ -7,7 +9,11 @@ export async function registerUser(
 ): Promise<ResRegisterDTO | undefined> {
   try {
     const result = await axios.post("http://localhost:3000/auth/register", dto);
-    return result.data;
+    console.log(jwtDecode(result.data));
+
+    localStorage.setItem("token", result.data);
+
+    return jwtDecode(result.data);
   } catch (error: any) {
     throw new Error(error.response.data.message);
   }
@@ -16,9 +22,11 @@ export async function registerUser(
 export async function loginUser(dto: loginDTO) {
   try {
     const result = await axios.post("http://localhost:3000/auth/login", dto);
-    console.log(result);
+    localStorage.setItem("token", result.data);
     return result.data;
   } catch (error: any) {
+    console.log(error);
+
     throw new Error(error.response.data.message);
   }
 }
