+++ backend/package.json
@@ -41,6 +41,7 @@
     "express": "^4.18.2",
     "express-fileupload": "^1.4.3",
     "jsonwebtoken": "^9.0.2",
+    "jwt-decode": "^4.0.0",
     "pg": "^8.11.3",
     "pg-pool": "^3.6.1",
     "postgres": "^3.4.3",
