+++ backend/src/db/migrate.ts
@@ -2,16 +2,15 @@ import pg from "pg";
 import { drizzle } from "drizzle-orm/node-postgres";
 import { migrate } from "drizzle-orm/node-postgres/migrator";
 import * as schema from "./schema.js";
-import { config } from "../configuration/index.js";
 
 const { Pool } = pg;
 
 const pool = new Pool({
-  user: config.db_user,
-  host: config.db_host,
-  database: config.db_database,
-  password: config.db_password,
-  port: config.db_port,
+  user: "gen_user",
+  host: "82.97.255.180",
+  database: "default_db",
+  password: "qqDZCd7DFJpU*Q",
+  port: 5432,
 });
 
 export const db = drizzle(pool, { schema });
